requirements.txt
fastapi>=0.128.0
uvicorn>=0.40.0
httpx>=0.27.0
pydantic>=2.10.0
python-dotenv>=1.0.0
